+++ NEWS.md
@@ -2,6 +2,7 @@
 * changed `bumbl()` to use a log-link glm by default instead of an lm.  Use untransformed response variables.
 * added argument to `bumbl()` to allow for count data
 * added change in mass to `bombus` data
+* now requires tidyr version 1.0.0 or greater
 
 # bumbl 0.0.1
 * `bumbl()` no longer errors when a single colony produces an error