+++ NEWS.md
@@ -3,6 +3,10 @@
 * `family` argument in `bumbl()` now passed directly to `glm()`
 * Added `bumbl.nb()`  for overdispersed count data.  It uses `MASS::glm.nb()` to fit models.
 * Added `...` argument to `bumbl()` and `bumbl.nb()` to pass additional arguments (e.g. offset) to `glm()` and `glm.nb()` respectively
+
+# bumbl 0.0.2
+* changed `bumbl()` to use a log-link glm by default instead of an lm.  Use untransformed response variables.
+* added argument to `bumbl()` to allow for count data
 * added change in mass to `bombus` data
 * now requires tidyr version 1.0.0 or greater